requirements.txt
Flask==2.0.3
python-telegram-bot==20.0
python-dotenv==0.19.2
werkzeug==2.0.2
